+++ Kernel/System/SysConfig.pm
@@ -10,6 +10,8 @@ package Kernel::System::SysConfig;
 
 use strict;
 use warnings;
+
+use Time::HiRes();
 use utf8;
 
 use Kernel::System::VariableCheck qw(:all);
@@ -3041,8 +3043,8 @@ sub ConfigurationListGet {
     my @ConfigurationList = $SysConfigDBObject->DefaultSettingListGet(
         Navigation               => $Param{Navigation},
         UserModificationPossible => $Param{TargetUserID} ? 1 : undef,
-        UserPreferencesGroup => $Param{UserPreferencesGroup} || undef,
-        IsInvisible => $Param{Invisible} ? undef : 0,
+        UserPreferencesGroup     => $Param{UserPreferencesGroup} || undef,
+        IsInvisible              => $Param{Invisible} ? undef : 0,
         %CategoryOptions,
     );
 
@@ -3696,6 +3698,23 @@ sub ConfigurationDeploySync {
     my $CurrentDeploymentID = $Kernel::OM->Get('Kernel::Config')->Get('CurrentDeploymentID') || 0;
 
     my $SysConfigDBObject = $Kernel::OM->Get('Kernel::System::SysConfig::DB');
+
+    # Check that all deployments are valid, but wait if there are deployments in add procedure
+    my $CleanupSuccess;
+    TRY:
+    for my $Try ( 1 .. 40 ) {
+        $CleanupSuccess = $SysConfigDBObject->DeploymentListCleanup();
+        last TRY if !$CleanupSuccess;
+        last TRY if $CleanupSuccess == 1;
+        sleep .5;
+    }
+    if ( $CleanupSuccess != 1 ) {
+        $Kernel::OM->Get('Kernel::System::Log')->Log(
+            Priority => 'error',
+            Message  => "There are invalid deployments in the database that could not be removed!",
+        );
+        return;
+    }
 
     my %LastDeployment = $SysConfigDBObject->DeploymentGetLast();
 
@@ -4692,9 +4711,9 @@ sub SettingsSet {
 
     # Deploy successfully updated settings.
     my %DeploymentResult = $Self->ConfigurationDeploy(
-        Comments => $Param{Comments} || '',
-        UserID   => $Param{UserID},
-        Force    => 1,
+        Comments      => $Param{Comments} || '',
+        UserID        => $Param{UserID},
+        Force         => 1,
         DirtySettings => \@DeploySettings
     );